requirements.txt
streamlit
txtai[similarity]
s3fs
firebase-admin
google-cloud-secret-manager
google-auth